requirements.txt
# --- CORE ---
langgraph
langchain-core
langchain-groq
langchain-community
langsmith

# --- BRAINS & MEMORY ---
semantic-router
fastembed
qdrant-client
mem0ai
instructor

# --- INFRA ---
pydantic
python-dotenv
sqlalchemy
requests
groq
streamlit
fastapi
uvicorn[standard]
sse-starlette
watchdog
pyperclip
sentence-transformers
